fix(globe): lift popup to fixed DOM (no flicker), fix night/realistic lighting (#47)
